Common:  - Prototype implementation of the path-based conditional signalling system  - Client-side rendering for signal lights changed
Germany:
 - Add Ne signs (1-5)
 - Bugfix: Ks signals now render Zs3 and Zs3v plate regardless of variant
 - Crossing gates

 Sweden:
  - Signal light aspects updated

+++ germany/src/main/java/venomized/mods/extendedsignals/de/signalling/HvCombinedSignalAspectCompositor.java
@@ -12,7 +12,7 @@ public record HvCombinedSignalAspectCompositor(SignalStateNode rawState,
      * @param states
      */
     @Override
-    public void applyAspect(long totalTicksForBlockEntity, SignalLighting states) {
+    public void applyAspect(float totalTicksForBlockEntity, SignalLighting states) {
         HvMainSignalAspect.interpret(rawState, direction)
                 .applyAspect(totalTicksForBlockEntity, states);
