+++ germany/src/main/java/venomized/mods/extendedsignals/de/signalling/HvMainSignalAspect.java
@@ -30,7 +30,7 @@ public enum HvMainSignalAspect implements IMainSignalAspect {
      * @param states
      */
     @Override
-    public void applyAspect(long totalTicksForBlockEntity, SignalLighting states) {
+    public void applyAspect(float totalTicksForBlockEntity, SignalLighting states) {
         for (String litLight : litLights) {
             states.powered(litLight);
         }
