+++ germany/src/main/java/venomized/mods/extendedsignals/de/signalling/KsMainSignalAspect.java
@@ -27,7 +27,7 @@ public enum KsMainSignalAspect implements IMainSignalAspect {
      * @param states
      */
     @Override
-    public void applyAspect(long totalTicksForBlockEntity, SignalLighting states) {
+    public void applyAspect(float totalTicksForBlockEntity, SignalLighting states) {
         for (String litLight : litLights) {
             states.powered(litLight);
         }
